requirements.txt
python==3.6.7
tensorflow-gpu==1.12.0
tensorflow-probability==0.5.0
matplotlib==3.1.0
colorcet==2.0.1
brokenaxes==0.3.1
sacred==0.7.5
pymongo==3.8.0
